Move opcodes to new module
